RF: Split options
