+++ fendermint/app/src/options/run.rs
@@ -1,0 +1,7 @@
+// Copyright 2022-2023 Protocol Labs
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use clap::Args;
+
+#[derive(Args, Debug)]
+pub struct RunArgs;
